refactor(tourism): 统一主键类型为Long并优化图片资源模型
- 将所有旅游模块主键和外键从Integer改为Long以匹配数据库BIGINT类型
- 在景区、场馆、活动、新闻等实体中增加cover_image_id字段，建立与图片资源的引用关系
- 重构图片服务接口，新增ensureByUrl和getImageMap方法用于图片资源统一管理
- 优化活动审核逻辑，在活动主表中增加audit_status字段，减少历史表查询
- 更新控制器和服务层接口，统一使用Long类型参数
- 删除临时文档文件，更新项目文档和论文内容

+++ cquet-tourism/src/main/java/cn/edu/cquet/tourism/controller/admin/TourismNewsController.java
@@ -55,7 +55,7 @@ public class TourismNewsController extends BaseController {
      * 入参：路径参数 `id`
      * 返回：新闻详情
      */
-    public Result getInfo(@PathVariable Integer id) {
+    public Result getInfo(@PathVariable Long id) {
         TourismNews news = tourismNewsService.getById(id);
         return success(news);
     }
@@ -96,7 +96,7 @@ public class TourismNewsController extends BaseController {
      * 删除
      * 入参：路径参数 `ids`
      */
-    public Result remove(@PathVariable List<Integer> ids) {
+    public Result remove(@PathVariable List<Long> ids) {
         // 将删除的结果转换为Result对象返回
         return toAjax(tourismNewsService.removeBatchByIds(ids));
     }
